Renamed ArtistsService to ArtistService, factored out all search-related methods into new classes ArtistSearchService and ArtistSearchServiceImpl
